+++ interop/kotlinx-metadata/src/main/kotlin/com/squareup/kotlinpoet/metadata/classinspectors/ReflectiveClassInspector.kt
@@ -73,6 +73,7 @@ import kotlin.metadata.jvm.syntheticMethodForAnnotations
 import kotlin.metadata.kind
 
 public class ReflectiveClassInspector private constructor(
+  private val lenient: Boolean,
   private val classLoader: ClassLoader?,
 ) : ClassInspector {
 
@@ -103,7 +104,7 @@ public class ReflectiveClassInspector private constructor(
       ?: error("No type element found for: $className.")
 
     val metadata = clazz.getAnnotation(Metadata::class.java)
-    return when (val kotlinClassMetadata = metadata.readKotlinClassMetadata()) {
+    return when (val kotlinClassMetadata = metadata.readKotlinClassMetadata(lenient)) {
       is KotlinClassMetadata.Class -> kotlinClassMetadata.kmClass
       is KotlinClassMetadata.FileFacade -> kotlinClassMetadata.kmPackage
       else -> TODO("Not implemented yet: ${kotlinClassMetadata.javaClass.simpleName}")
@@ -249,7 +250,7 @@ public class ReflectiveClassInspector private constructor(
         // class.
         null
       } else {
-        enumEntry.javaClass.getAnnotation(Metadata::class.java)?.toKmClass()
+        enumEntry.javaClass.getAnnotation(Metadata::class.java)?.toKmClass(lenient)
       },
       annotations = clazz.getField(enumEntry.name).annotationSpecs(),
     )
@@ -540,9 +541,12 @@ public class ReflectiveClassInspector private constructor(
   }
 
   public companion object {
+    /**
+     * @param lenient see docs on [KotlinClassMetadata.readStrict] and [KotlinClassMetadata.readLenient] for more details.
+     */
     @JvmStatic
-    public fun create(classLoader: ClassLoader? = null): ClassInspector {
-      return ReflectiveClassInspector(classLoader)
+    public fun create(lenient: Boolean, classLoader: ClassLoader? = null): ClassInspector {
+      return ReflectiveClassInspector(lenient, classLoader)
     }
 
     private val Class<*>.descriptor: String