+++ interop/ksp/test-processor/src/main/kotlin/com/squareup/kotlinpoet/ksp/test/processor/exampleAnnotations.kt
@@ -81,3 +81,4 @@ enum class AnnotationEnumValue {
 }
 
 annotation class AnnotationWithVararg(val simpleArg: Int, vararg val args: String)
+annotation class AnnotationWithTypeArgs<T, R>